+++ docs/class-diagram.plantuml
@@ -76,6 +76,37 @@ package io.singularitynet.sdk.registry {
 
 }
 
+package io.singularitynet.sdk.daemon {
+    
+    interface DaemonConnection {
+        <T> T getGrpcStub(Function<Channel, T> constructor);
+        void setClientCallsInterceptor(ClientInterceptor interceptor);
+        String getEndpointGroupName();
+        void shutdownNow();
+    }
+
+    class RandomEndpointDaemonConnection
+    DaemonConnection <|.. RandomEndpointDaemonConnection
+    RandomEndpointDaemonConnection o-- MetadataProvider
+
+    class PaymentChannelStateService {
+        PaymentChannelStateReply getChannelState(BigInteger channelId);
+    }
+    PaymentChannelStateService o-- Identity
+    PaymentChannelStateService o-- Ethereum
+    PaymentChannelStateService -- DaemonConnection
+
+    class FreeCallStateService {
+        long getFreeCallsAvailable(String dappUserId, String freeCallToken,
+                BigInteger tokenExpirationBlock, Identity signer);
+    }
+    FreeCallStateService o-- Ethereum
+    FreeCallStateService o-- MetadataProvider
+    FreeCallStateService o-- DaemonConnection
+
+}
+
+
 package io.singularitynet.sdk.mpe {
 
     class MultiPartyEscrowContract {
@@ -108,6 +139,11 @@ package io.singularitynet.sdk.mpe {
     BlockchainPaymentChannelManager <|.. MpePaymentChannelManager
     MpePaymentChannelManager o-- MultiPartyEscrowContract
 
+    class AskDaemonFirstPaymentChannelProvider
+    PaymentChannelStateProvider <|.. AskDaemonFirstPaymentChannelProvider
+    AskDaemonFirstPaymentChannelProvider o-- MultiPartyEscrowContract
+    AskDaemonFirstPaymentChannelProvider *-- PaymentChannelStateService
+
 }
 
 package io.singularitynet.sdk.freecall {
@@ -116,42 +152,6 @@ package io.singularitynet.sdk.freecall {
     Payment ()- FreeCallPayment 
 
 }
-
-package io.singularitynet.sdk.daemon {
-    
-    interface DaemonConnection {
-        <T> T getGrpcStub(Function<Channel, T> constructor);
-        void setClientCallsInterceptor(ClientInterceptor interceptor);
-        String getEndpointGroupName();
-        void shutdownNow();
-    }
-
-    class RandomEndpointDaemonConnection
-    DaemonConnection <|.. RandomEndpointDaemonConnection
-    RandomEndpointDaemonConnection o-- MetadataProvider
-
-    class PaymentChannelStateService {
-        PaymentChannelStateReply getChannelState(BigInteger channelId);
-    }
-    PaymentChannelStateService o-- Identity
-    PaymentChannelStateService o-- Ethereum
-    PaymentChannelStateService -- DaemonConnection
-
-    class AskDaemonFirstPaymentChannelProvider
-    PaymentChannelStateProvider ()- AskDaemonFirstPaymentChannelProvider
-    AskDaemonFirstPaymentChannelProvider o-- MultiPartyEscrowContract
-    AskDaemonFirstPaymentChannelProvider *-- PaymentChannelStateService
-
-    class FreeCallStateService {
-        long getFreeCallsAvailable(String dappUserId, String freeCallToken,
-                BigInteger tokenExpirationBlock, Identity signer);
-    }
-    FreeCallStateService o-- Ethereum
-    FreeCallStateService o-- MetadataProvider
-    FreeCallStateService o-- DaemonConnection
-
-}
-
 
 package io.singularitynet.sdk.client {
